test: fixtures and integration matrix on the new ABI
- error_plugin allocates via HostInterface passed in args (no extern
  symbol); csharp fixture uses (uint)AbiErrorCode + correct bundle id
- manifest ids in inline test manifests computed via bundle_id(name)
- stale HOST_CONTRACT_* env emission removed from build.rs (fossil of
  the logger example migrated in d33d1db1; zero consumers)
- cpp hot_reload_test trimmed to real ReloadPhase coverage after
  runtime_config.hpp fossil deletion; deno host fixture on canonical
  find/resolveGuestContract names; js fixture registers with version

+++ tests/fixtures/csharp_plugin/Plugin.cs
@@ -77,7 +77,7 @@ public static class Plugin
             var outPtr = (uint*)result;
             *outPtr = addArgs->A + addArgs->B;
         }
-        return new AbiError { Code = AbiErrorCode.Ok };
+        return new AbiError { Code = (uint)AbiErrorCode.Ok };
     }
 
     [UnmanagedCallersOnly]
@@ -89,7 +89,7 @@ public static class Plugin
             var outPtr = (uint*)result;
             *outPtr = addArgs->A + addArgs->B;
         }
-        return new AbiError { Code = AbiErrorCode.Ok };
+        return new AbiError { Code = (uint)AbiErrorCode.Ok };
     }
 
     [UnmanagedCallersOnly]
@@ -103,13 +103,13 @@ public static class Plugin
                 *outPtr = new StringView { Ptr = (nint)ptr, Len = (nuint)s_versionBytes.Length };
             }
         }
-        return new AbiError { Code = AbiErrorCode.Ok };
+        return new AbiError { Code = (uint)AbiErrorCode.Ok };
     }
 
     [UnmanagedCallersOnly]
     public static AbiError Reset(nint args, nint result)
     {
-        return new AbiError { Code = AbiErrorCode.Ok };
+        return new AbiError { Code = (uint)AbiErrorCode.Ok };
     }
 
     [UnmanagedCallersOnly(EntryPoint = "PolyplugInit")]
